pyproject.toml
[project]
name = "rhizome"
version = "0.1.0"
description = "A rhizomatic Zettelkasten system for atomizing and connecting notes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "ollama>=0.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
]

[project.scripts]
rhizome = "rhizome.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
